When GC'ing a native method don't try to free code
A native method's code pointer is either NULL or, for a JNI
native method, points to the native function.  In the former
case freeing is a no-op, but in the latter we will try to
free a non-malloced pointer.  This oversight has existed
ever since the class GC code was added.  Why it hasn't caused
a problem is a mystery - either no one has reported it, or
I've been lucky and no class with a native method has ever
been freed...

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>