Fix non-direct interpreter invokespecial super-class check
Just over a week ago I fixed the invokespecial check that
determined if it was a super-class invocation rather than a
non-virtual invocation.  I did this for the direct-interpreter
variants.  This checkin does the same for the non-direct variants.

Note this fix is not required for bootstrapping of IcedTea3.  It
is however required with the HEAD of OpenJDK 8 (as of 2-3 weeks
ago).  As suspected, the issues were unrelated.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>