JSR 335: implement "bridge" methods for Mirandas
Miranda methods are methods created by the VM to represent
unimplented interface methods.

In the past, Miranda methods were copies of the interface
method - as they were abstract, we'll get an abstract method
error.

With JSR 335, they could also be a default interface method.
Copying the interface method in this case required the method
in the implementing class to have the interface's class
(previously, all methods defined in a class had that class as
their class).

This was nasty, and also required a hack in the GC.  But the real
problem I only realised later.  With JSR 335, an interface method
can be executed directly (via invoke special).  This means
that there is the possibility that when we're copying the interface
method it is preparing - in this case, the copied method will
hang on preparation (it will wait for it to be prepared, but
the actual method being prepared is the original, so its state
will never change).

We can also copy a prepared method.  With inlining, this will end
up sharing pointers to inlining information, which will become
inconsistent.

With this change, Miranda methods simply "point" to the unimplemented
interface method.  Executing the Miranda method sets up the frame -
the Miranda method code then transfers execution to the interface
method (i.e. it acts as a bridge).

The advantage of this is that if multiple classes end up with the
default we only have one copy of the default.  Previously, we
would end up with multiple copies, and multiple sets of prepared
code.

Performance-wise, we end up with an extra opcode (the body of the
Miranda) that does the bridge.  Compared to the cost of creating
the frame, etc. this is minimal.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>