JSR 335: remove GC hack
Initial support for default interface methods required
that the Miranda had the interface's class.  This broke
a GC assumption that all methods defined by a class had
the class as their class (a reasonable assumption).  So
a temporary hack was introduced.  Now, with bridge methods,
this hack can be removed.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>