FreeClassData: adjust method count for Miranda methods
The size of the method extra attributes is equal to the number
of methods declared in the class when the class was defined.
However, the methods may be extended at link time to include
Miranda methods (as these do not have extra attributes, the
extra attributes are not extended).  When freeing the method
extra attributes we therefore need to adjust the methods count
to get the original, non-Miranda value.