JSR 292: invokedynamic is 5 bytes in length
For some reason when implementing invokedynamic I assumed it
was 3 bytes in length, like most of the invoke bytecodes
(1 byte bytecode followed by a 2 byte constant-pool index).
This worked, and I've had no reason to question it.  However,
idly reading the spec,  I noticed it is actually 5 bytes
in length (same as an invokeinterface).

The reason it has worked when treated as 3 bytes is that the
other 2 bytes must be zero.  This meant they appeared as two NOP
instructions.

This change fixes it for the direct-threaded/inlining interpreter,
when the bytecodes are rewritten on method preparation.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>