JSR 292: invokedynamic is 5 bytes - non-direct interp
Previous checkin explains that invokedymamic was implemented
as being 3 bytes in length when in fact it is 5 (as other bytes
were 0, they appeared as NOPs so everything still worked).

Previous checkin fixed it for the direct/inlining interpreter -
this fixes it for the old non-direct interpreter variants.  This
isn't used in "production" code, but I often build these variants as
they are easier to debug.

Note, the main change is the two single line changes in interp.c.  The
rest of the changes rearrange some of the internal bytecodes, so that
both of the 5-byte invoke bytecodes can be detected using a single
if (we change it from an == to a >=).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>