JSR 335: java.lang.invoke.MagicLambdaImpl
While implementing JSR 292 with OpenJDK 8, I needed to
handle java.lang.invoke.MagicLambdaImpl specially during
access checks (similar to sun.reflect.MagicAccessorImpl).

Now, Lamda expressions and JSR 292 are closely related in
OpenJDK 8, so I lumped this in with the JSR 292 code (ifdefed
with JSR 292).

However, IcedTea 2.4.0 is based on OpenJDK 7 update 40, which
has the backport of JSR 292 from OpenJDK 8.  However, it doesn't
have Lamda expressions (as these are part of OpenJDK 8).

I hoped JamVM's implementation of JSR 292 would just work.  It
didn't, because the MagicLamdaImpl stuff is missing.  To fix
this, I've put it under a new macro, JSR 335, which is defined
along with JSR 292 for OpenJDK 8, but isn't defined for IcedTea
7.  With this, the JSR 292 implementation works with OpenJDK 7.
To be honest, I think this is a pretty good result.  JamVM now
supports JSR 292 in both OpenJDK 8 and OpenJDK 7!

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>