JSR 292: enable for OpenJDK 7/IcedTea 2
IcedTea 2.4.0 is now based on OpenJDK update 40, which
contains the backport of JSR 292 from OpenJDK 8.

With a minor tweak (see previous checkins), JamVM's support
for JSR 292 with OpenJDK 8 works with IcedTea 2.4.0.  As IcedTea
is the primary upstream client for JamVM/OpenJDK, this change
enables JSR 292 for OpenJDK 7 in addition to OpenJDK 8.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>