JSR 335: handle interfaces when resolving a method
With JSR 335, method resolution can result in a resolved
interface class (previously this was an error).  In this
case we must search not just the interface, but also the
interfaces which it extends.  We do not need to search
the implemented interfaces for non-interface classes, as
all methods must be present on the class (implemented or
as VM-generated Miranda methods).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>