Add support for CallerSensitive
Some methods are sensitive to where they are called from.
In particular, methods which obtain information from the stack.
Previously, stack layout assumptions were hard-coded but this
is fragile, and hard to ensure with method handle adapters.
The CallerSensitive annotation marks such methods removing
the need for hard-coding.

This checkin adds support for the CallerSensitive annotation.
This is used to set the CALLER_SENSITIVE flag when constructing
MemberNames.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>