GC: Minor performance improvement
During compaction, when threading class data, we know the new
address of the class; this is used to directly update the class
reference within the method and field data (no point in threading
and then unthreading them).  However, we don't need to do anything
if the class isn't actually moving (i.e. the new address is the
same as the old address).  This will happen if a reference to
the class has been found conservatively (references to other
objects from within the class data still need to be threaded, but
the class itself won't be moved).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>