GNU Classpath: fix threading issue in lookupLoadedDlls0
LookupLoadedDlls0 searches the loaded shared libraries for
a symbol (i.e. a native method).  The loaded shared libraries
are contained within a hashtable.  Libraries are added to
the hashtable in resolveDll, which may require the table to
be resized, etc.

To ensure safe access via multiple threads, the hashtable must
be locked in both resolveDll and lookupLoadedDlls0.  However,
locking the table in lookupLoadedDlls0 occasionally lead to
deadlocks.  As a temporary "fix" I removed the locking in
lookupLoadedDlls0, intending to fix it properly (in reality,
library loading occurs rarely), and it's been like it for years.
However, two users of GNU Classpath have reported it independently
fairly recently (within the last year).

Tracking the original deadlock down shows the cause is internal
locking within the native dynamic linking loader.  Both resolveDll
and lookupLoadedDlls0 call dlysm.  In lookupLoadedDlls0 this is
done while the hashtable lock is held (or it was before the lock
was removed). The problem is, hashtables need to be scanned by
GC, so hashtables have suspension disabled while the lock is held
(fast disable, so the suspension code waits for them to self
suspend).

The "deadlock" occurs as follows:

* thread A calls dlsym in resolveDll, and obtains internal lock

* thread B initiates GC, and suspends thread A while it is holding the
  internal lock

* thread C calls dlsym in lookupLoadedDlls0 and tries to obtain the
  internal lock

As thread A holds the internal lock, thread C blocks.  Thread B
then tries to suspend C, sees that it has suspension disabled,
and waits for it to self suspend.  But thread C will never self
suspend as it is blocked by thread A which is suspended, so the
suspension code hangs.

The problem is that lookupLoadedDlls0 breaks the rule that
fastDisable/fastEnable sections can never block.  To prevent this,
a thread in resolveDll cannot be allowed to be suspended holding
the internal lock, i.e. it must also have suspension disabled
while calling dlsym.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>