JSR 292: correct intrinsic cache COMPARE function
When the intrinsic method ID was moved to the method state field
from the flags field, the hashtable HASH and COMPARE functions
needed updating.  However, only the HASH function was updated.
This hasn't caused a problem as the hash value is unique per
intrinsic (two different intrinsics can't get the same hash value).
The additional checks on the separate fields when the hash value
matched was simply a precaution.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>