JSR 308: update to match OpenJDK changes
Initial support for JSR 308 in OpenJDK stored the type annotations
in an additional instance variable that had to be separately
initialised when the VM created a reflection object.  This was
wasteful as type annotations are rarely used.  The instance
field have now been removed, and the type annotation data is
retrieved lazily via additional JVM_xxx functions.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>