GNU Classpath: race-condition in thread detach
Nasty race condition in thread detach introduced by the
classlib restructuring.  When the Java thread object is marked
as terminated, the VMThread reference is set to NULL.  The
problem is that the native thread structure is tied to the
VMThread, and is freed when the VMThread becomes unreachable.

This means there is a small race-condition between setting the
VMThread reference to NULL, and grabbing the thread lock.  If
a GC were to occur, and the only reference to the VMThread had
been from the Java thread (likely) then the native thread structure
will be freed (and there are several references to it after this
point).

The reason this crept in is because much of the details had to
be abstracted to support OpenJDK and GNU Classpath (inside
detachThread there is only a reference to the Java thread).

The fix is to maintain a reference to the VMThread object to
prevent the thread structure being freed.  This has been abstracted
into a "keep_alive" object returned by the classlib mark thread
terminated function.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>