JSR 292: bootstrap method args error handling
When I wrote the bootstrap method args code I forgot that
resolveSingleConstant does not return NULL on error.  Previously,
the way to tell if an error had occurred in one of the resolveXXX
functions was to check exceptionOccurred.  However, this is
(potentially) slow, so the code was changed so that in most
cases NULL was also returned.  However, this doesn't apply to
resolveSingleConstant, as the return value is an opaque uintptr_t,
and it can potentially return 0 in a non-error situation (e.g.
resolving a CONSTANT_Integer).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>