JSR 292: MethodHandle constant-pool refs may be resolved
A method handle constant-pool entry contains the index of the
referenced entry in the constant-pool and the reference kind (the
bytecode behaviour).  The referenced entry can be a field, a method
or an interface method.  The class, name and type of the entry is
passed up to the JSR 292 runtime, which links it and returns a
MethodHandle.

For some unknown reason, when I wrote the code initially, I
resolved the class, and peeked the name and type from the referenced
entry (although there may be different entries, they're all stored
the same).  I was probably following the spec too closely.  The
Java code also simply calls back into the VM to resolve the entry
so maybe I was avoiding doing resolution twice.

However, I forgot the entry may have been resolved from a different
path.  In this case, peeking a resolved entry gives wrong indexes,
and the code fails in a random way.

Fix is to resolve the entry via the correct resolveXXX function and
then obtain the class, name and type from the resolved entry.  The
reference kind is used to determine the correct resolveXXX function to
call.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>