Fix race-condition in constant-pool resolution code
The constant-pool resolution code is lock-less.  Instead
it relies on the correct ordering of reads/writes, which it
ensures using memory-barriers.

Recently I noticed an intermittent crash with Dacapo xalan benchmark
which I tracked down to the resolution code.  Recent versions of gcc
are re-arranging some loads which break the required ordering.

For example, in the following code, the load for the re-check of
the CP_TYPE must be done after the load for the name_idx (because
if the type is still the same, we know the name_idx is correct as
it was read before).

        case CONSTANT_Class: {
            char *classname;
            int name_idx = CP_CLASS(cp, cp_index);

            if(CP_TYPE(cp, cp_index) != CONSTANT_Class)
                goto retry;

However, gcc had re-ordered the loads so that the load for the name_idx
was done after that for the type.  This means there is a small race-
condition between checking the type and loading the name_idx where
the entry could be resolved by another thread.  In this case, the loaded
idx would be wrong.

Simple fix is to insert an extra memory barrier which prevents the
re-ordering.  Although a compile barrier would also do this, a memory
barrier has always been missing here, as some memory architectures
may re-order the loads even if they were executed in the correct order.