OpenJDK: Fix Local/Anonymous class checks
According to JVM spec 4.8.6 a class must have an EnclosingMethod
attribute if and only if it is a local or an anonymous class.

In constrast, a nested or inner class has a declaring class
specified in the InnerClasses attribute.

The enclosing class is therefore either the enclosing class from
the EnclosingMethod, or the declaring class from the InnerClasses
attribute.

When implementing this with GNU Classpath, I handled it by simply
setting the enclosing class to the declaring class when parsing the
InnerClasses attribute.  GetEnclosingClass then simply returned the
enclosing class (which could be the declaring class).

When implementing GetEnclosingMethodInfo on OpenJDK, it simply
returned the enclosing class and any enclosing method.  Unfortunately,
OpenJDK uses the presence of MethodInfo to tell if a class is a local
or anonymous class (as according to the spec, only these classes have
an EnclosingMethod attribute).  But because enclosing class is
set to the declaring class, nested or inner classes appear to have
MethodInfo, and therefore appear to be a local/anonymous class.
With GNU Classpath/JamVM, the anonymous/local class check
is implemented in a different way, and doesn't use the absence
of an enclosing class.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>