JSR 292: Fix CallStaticInitOrder jtreg test
A recent change to resolveMethodHandle made it call the resolveXXX
functions rather than peeking the referenced entry (as it may
already be resolved).  This had the side-effect of initialising
the method's defining class.  This is because resolveMethod
always initialised the class.  It previously was only used when
executing an invoke bytecode.  For static invokes, initing is required,
and for non-static invokes it was a no-op.

Fixed by moving the call to initClass from resolveMethod to the
invoke static rewriter.  This means it's only initialised when we
first execute the bytecode.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>