Do not initialise the class when resolving a field
A recent commit changed resolution of methods so that the declaring
class was no longer initialised - this was moved into the interpreter
and only done on first execution of a static method (as per the spec).
A similar issue exists with fields, although the failing jtreg method
handle test didn't fail (as it doesn't handle fields).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>