Optimise ClassBlock structure
I've known for some time that the ClassBlock structure, due
to being developed initially on a 32-bit machine contained
some redundant padding on a 64-bit machine (because of a mixture
of ints and pointers).

Rearrange structure to pack similar sized fields together to
minimise padding.  This saves 24 bytes on a 64-bit machine.

The rationale for doing this is to hide the extra field
host_class - this is redundant for most classes.  Best solution
is to develop a way to only add this for anonymous classes, but
for now, reducing the size of ClassBlock means the addition
doesn't add to the overall size.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>