Further optimisation of ClassBlock
The space removed in the previous change was mostly applicable
to 64-bit machines. However, several fields are mutually
exclusive between array and non-array classes.  Put these into
unions.  This overlaps two fields which saves 8 bytes on a
32-bit machine.  On a 64-bit machine, it also saves 8 bytes
as the overlapped int field is replaced by padding so it doesn't
reduce the overall size.

This is the optimal ClassBlock structure.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>