Explicitly initialise java.lang.Class
Another consequence of the change to class initialisation during resolution
is that with OpenJDK 7 java.lang.Class ends up being used non-statically
before initialisation.  Specifically, a native method is called, and because
the class hasn't been initialised, the native methods haven't been registered
and we get an UnsatisfiedLinkError.

Fix is to explicitly initialise java.lang.Class at an appropriate place
during VM initialisation (normal classes are ensured to be initialised
before creating an instance, but class objects are handled specially
within the VM).

Strangely, the problem didn't show up with OpenJDK 8 (presumably,
somewhere in the initialisation a static use is made first of
java.lang.Class but I haven't checked) .

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>