OpenJDK: use _FILE_OFFSET_BITS rather than _LARGEFILE64_SOURCE
Previously, to get 64-bit file support on glibc (e.g. Linux) the
_LARGEFILE64_SOURCE macro was used.  On 32-bit systems this provided
an additional set of functions ending in "64", e.g. lseek64 for
to lseek, and an additional offset type (off64_t).  On 64-bit
systems, the additional interface is identical to the old one.

Unfortunately, these do not exist on MacOS X - instead, lseek, etc.
are 64-bit by default, as on 64-bit Linux.  However, with glibc
an additional macro can be used - _FILE_OFFSET_BITS.  Defining this
to 64 replaces the old interface with a 64-bit interface on
32-bit systems.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>