JSR 292: add host_class to classlib GC refs
The CLASSLIB_CLASSBLOCK_REFS_DO macro enables classlib specific
fields within the classblock to be handled during GC.  Add host_class
to it, as the host_class may move due to compaction, and the reference
must be updated.  Strictly, class refs such as an array class's
element class do not need explicit marking as it will be marked via
the classloader.  The host_class, however, does not have this, and
so must also be explicitly marked.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>