Make getting size of physical memory OS dependent
During the port to OpenJDK I made the default minimum and maximum
heap sizes based on the size of physical memory.  At the time I was
only interested in Linux and it appears the code is not portable
(at least it doesn't work on MacOS X).

Move the function into the OS layer.  For now it is only implemented
on Linux - the other platforms return 0, i.e. not supported.  In
this case a fixed default minimum and maximum is used (as previously).
This means JamVM will compile and work on all platforms.  The
improved default sizes can then be rolled out on other platforms
later.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>