Allow -verbose in JNI init args
Very long running issue where -verbose wasn't accepted as a
valid argument when creating the VM via JNI.

When implementing the JNI invocation interface spec, I read that
-verbose must be followed by 1 or more comma separated names (class,
gc or jni).  I assumed that -verbose on its own was invalid.

As JamVM with GNU Classpath used its own launcher this was never an
issue.  However, when JamVM is used with OpenJDK, it is launched via
JNI by the OpenJDK launcher.  Specifying -verbose to the launcher
shows that it is passed directly to the JNI invocation API.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>