JSR 292: missing ifdefs on indirect interpreter
The handler for OPC_INVOKEDYNAMIC should only be defined when JSR 292
is enabled.  This wasn't the case for the indirect interpreter (threaded
or switch).  It hasn't been noticed as the indirect-threaded interpreter
is not built by default for any platform.  It is still supported as it
is useful for easier debugging.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>