JSR 292: Fix RepetitiveLambdaSerialization jtreg test
This took some tracking down.  Reflection uses native code to do an
invoke until a threshold is reached.  After this, a bytecode invoker
is generated.  This invoker needs a reference to the "method to be
invoked" class.  To do this, it generates a symbolic reference in the
constant pool.

So far so good.  The problem is, if the method to be invoked is in
an anonymous class (i.e. a lambda) the class can't be resolved
symbolically (as it is anonymous, it does not appear in any class-
loader tables).  The test therefore fails with an undefined class
error.

The solution in OpenJDK was to stop proxy generation for anonymous
classes.  To do this, it recognises the name of an anonymous class
(specifically, it must end in /<num>).  Now, anonymous classes are
created by the VM.  This test matches the names created by Hotspot,
but when implementing this in JamVM I didn't bother to append a unique
number.  So the test fails with JamVM.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>