Make ClassBlock padding dependent on OpenJDK version
The JamVM ClassBlock structure is embedded inside the java.lang.Class
object instance data (so given a pointer to a java.lang.Class instance,
the internal structure can be obtained simply by adding on an offset).

However, to do this the structure must be padded, so that when it is
overlaid onto the instance data none of the fields overwrites real
instance data.  Obviously, this padding is dependent on the number and
type of instance variables in java.lang.Class.

The padding value was initially determined for OpenJDK 6, and it
has "worked" ever since on OpenJDK 7 and 8.   However, while investigating
a bug with OpenJDK 8, I noticed that most of the padding was empty.

This lead to an investigation of what padding was actually required.
It turns out that the size of each class object has been reduced
quite substantially on OpenJDK 8, allowing the padding to be shrunk.

More worryingly, the investigation also showed that the number of fields
increased between OpenJDK 6 and OpenJDK 7.  As the size was padded
to a multiple of uintptr_t, the padding was still OK on 64-bit machines.
However, on 32-bit machines, the padding was 4 bytes too small.  I am
somewhat surprised that this has not caused any problems (e.g. on ARM).