Update GNU Classpath ClassBlock padding
The previous checkin made the ClassBlock padding dependent on
OpenJDK version.  As part of this checkin, the padding was changed
from a uintptr_t to char.  The use of uintptr_t was a hangover
from the days when instance data was based on reference sized
"slots" (similar to the operand stack).  This hasn't been the case
since JamVM was properly ported to 64-bit systems, and instance
data was laid out based on field size.

As padding is now char sized, the padding value for GNU Classpath
needs to be updated (size of 4 reference fields).

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>