Check ClassBlock padding at VM initialisation
The ClassBlock structure is padded so that it doesn't overlap
java.lang.Class instance data when overlaid onto a java.lang.Class
object.  The padding value is determined statically based on the
number and size of java.lang.Class instance variables.

If the padding is too small, corruption will occur.  However, in
the past there was no check to ensure that the static padding
value was sufficient.  This has never caused a problem (as far
as I'm aware).  However, as detailed in a previous commit, investigation
of padding sizes for OpenJDK revealed that with OpenJDK 7 on 32-bit
machines the padding was actually 4 bytes too small.

This checkin adds a check during VM initialisation that the padding
is at least as large as the java.lang.Class instance data.  If it
isn't, the VM fails initialisation.  It should be noted that this
check failed on 32-bit ARM with OpenJDK 7 with the old padding value.
This indicates the check works.  Obviously, it passes with the new
OpenJDK dependent padding values.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>