OpenJDK: Use correct thread group function when thread dies
When a thread terminates, the VM must call a method on the thread's
thread group to notify it of the thread death.  Under GNU Classpath,
the function was "removeThread".  When I implemented it for OpenJDK,
I found the function "remove" which seemed to be the same.  However,
this doesn't handle correctly daemon threads.  The correct function
is instead "threadTerminated" (which calls "remove").

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>