JSR 292: invokedynamic resolution cleanup 1
As discussed in the previous checkin, invokedynamic should be resolved
per callsite, not per callsite specifier.  This is nasty because it
breaks many of the existing resolution idioms - e.g. it requires multiple
entries to be cached per constant pool entry.  Anyway, previous checkin
fixed it so it worked.  This is the first of several cleanups:

1) Invokedynamic rewriting must be thread-safe.  Racing threads must see
   the same resolved callsite

2) The return results from findInvokeDynamicInvoker must be GC-safe.
   Previously there was a window before the PolyMethodBlock was stored
   into the constant-pool where the appendix in the pmb was potentially
   unreachable.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>