Minor interpreter tidy-up
The invokedynamic fix involved splitting out the OPCODE_CHANGE
part of WITH_OPCODE_CHANGE... into a seperate macro so it could be
used independently.  This was done for the inlining interpreter,
but the macro(s) are common to the direct interpreter.  Moved
it to the common header file, and removed it from the others.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>