JSR 292: invokedynamic resolution fix for old interp
The fix for the invokedynamic resolution issue needs extra changes
to work with the old interpreter.  This is because there isn't enough
room in the bytecode stream to pull in the reference to the specific
callsite resolved for that instruction.  Instead, we can only store
and index into the linked list (it must be a linked-list to enable it
to be accessed lock-less).

This initial implementation isn't particularly efficient as it needs
to search the linked-list on each invocation.  However, the list is
expected to the be small, and the old interpreter is only retained for
easier debugging.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>