JSR 292: invokedynamic resolution fix for old interp (v2)
As mentioned in the previous commit, the initial implementation wasn't
particularly efficient as it needs to search the linked-list on each
invocation.  The expectation is that the list would mostly have only 1
element.  On further testing with nashorn this is the case, but I also
saw lists created that were 4 and 6 elements long.  Not massive, but
still large enough to no longer be happy with the first implementation.
The new one introduces a cache, so repeated invocations of the same
callsite will not need searching.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>