JSR 292: Free callsite linked-list on class GC
The linked-list of resolved callsites must be walked and freed
when the class is GC'ed.  Note, normally you need to be careful
to get the next pointer _before_ you free the current block.  We
don't need to do this here as the blocks are added to the pending
free list and freed at the end of GC.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>