JSR 292: findInvokeDynamicInvoker/resolveCallSite tidy-up
In a similar way to the previous check-in, we also need to
be careful in findInvokeDynamicInvoker to ensure the appendix
is reachable.  Likewise in resolveCallSite we are allocating
memory and must prevent a leak when threads are racing.  I realised
this when writing the code (this is what lead me to realise the
problem in resolvePolyMethod).  However, I returned the appendix
and passed the invoker as a parameter reference (in fact, I had
to return the appendix box as the appendix can be null, i.e.
indistinguishable from an error).  When I fixed resolvePolyMethod,
I instead passed the appendix via the pointer-parameter.  This
is just as safe, as the reference will be saved into the caller
stack frame (i.e. reachable).  However, it looks neater having
functions called findXXXInvoker return the invoker rather than an
object!

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>