Fix building unoptimised
GNU Classpath and OpenJDK without JSR292 have no injected fields.
Therefore classlibInjectFieldsCount is a constant 0.  This means
in the following code classlibFillInInjectedFields is normally
optimised out, and we don't need to provide it:

if(injected_fields_count != 0)
    classlibFillInInjectedFields(...)

However, when building without optimisation it isn't removed and
we get an undefined symbol.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>