JSR 292: show full anonymous/lamdba class name in stack trace
An anonymous class (used for lambda classes) has its initial location
appended to the name to make it unique (xxx/num).  This, however, is
not stored internally - it is appended when the class-library requests
the name.  This was sufficient to fix the problem where the class-library
didn't recognise the class as anonymous (based on the /num).  I forgot
however, that classnames are also returned to the class-library via a
stack trace (they are filled in by the VM).  So in the following program
we get a different name in the stacktrace as that printed.

public class Lambda {
    public static void main(String[] args) {
        Runnable r = () -> { Object o = null; o.toString(); };
        System.out.println(r);
        r.run();
    }
}

Added a new classlib function to return the external name, which is now
used when filling in the stack trace.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>