OpenJDK: thread termination - use remove on OpenJDK 6
A couple of weeks ago I made a change to the method called by JamVM
on thread termination.  This method removes the thread from the thread
group, and was changed from "remove" to "threadTerminated".  The
reason for the change was to fix a jtreg test.  At the time I could
not understand why I used remove.  However, I know now!  It doesn't
exist on OpenJDK 6...

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>