OpenJDK: recognise extra management boolean attributes
The management implementation is a stub and doesn't do anything.
However, it does recognise the full set of attributes - unrecognised
ones cause a trace to be printed (idea being it should be checked).
For a long time, I've noticed a log message is printed when running
jBidwatcher.  Turns out there are some missing boolean attributes.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>