Fix JamVM with OpenJDK 8
Coming out of a 5 month hibernation I see JamVM is broken with
the latest update of OpenJDK 8.  The fix is trivial - an extra
reference field has been added to java.lang.Class, and so the
ClassBlock padding needs to be expanded.

If it wasn't for the copyright update this would be a 1 byte
fix!

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>